Change ingredient functionality. Can create ingredient, but does not display list

+++ praktinis_darbas/src/main/java/lt/techin/dto/shoppingListItem/ShoppingListItemResponseDTO.java
@@ -1,13 +1,14 @@
 package lt.techin.dto.shoppingListItem;
 
-import lt.techin.dto.ingredientCategory.IngredientCategoryResponseDTO;
+import lt.techin.dto.ingredient.IngredientCategoryResponseDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ShoppingListItemResponseDTO(
         Long id,
         String ingredientName,
-        Double quantity,
+        BigDecimal quantity,
         String unit,
         List<IngredientCategoryResponseDTO> ingredientCategory
 ) {
